+++ app/src/main/java/com/emplk/realestatemanager/domain/currency_rate/CurrencyRateEntity.kt
@@ -1,8 +1,9 @@
 package com.emplk.realestatemanager.domain.currency_rate
 
+import java.math.BigDecimal
 import java.time.LocalDate
 
 data class CurrencyRateEntity(
-    val usdToEuroRate: Double,
+    val usdToEuroRate: BigDecimal,
     val lastUpdatedDate: LocalDate,
 )
