+++ app/src/main/java/com/emplk/realestatemanager/domain/currency_rate/CurrencyRateWrapper.kt
@@ -1,6 +1,8 @@
 package com.emplk.realestatemanager.domain.currency_rate
 
+import java.math.BigDecimal
+
 sealed class CurrencyRateWrapper {
     data class Success(val currencyRateEntity: CurrencyRateEntity) : CurrencyRateWrapper()
-    data class Error(val fallbackUsToEuroRate: Double) : CurrencyRateWrapper()
+    data class Error(val fallbackUsToEuroRate: BigDecimal) : CurrencyRateWrapper()
 }
